code reorganization: extracted PythonLogHandler into separate files
