+++ src/client/CMakeLists.txt
@@ -1,11 +1,11 @@
 # Copyright (c) 2012-2020 Esri R&D Center Zurich
- 
+
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
- 
+
 #   http://www.apache.org/licenses/LICENSE-2.0
- 
+
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -25,7 +25,8 @@ pybind11_add_module(${CLIENT_TARGET} MODULE
 		utils.cpp
 		wrap.cpp
 		PyCallbacks.cpp
-		PRTContext.cpp)
+		PRTContext.cpp
+		PythonLogHandler.cpp)
 
 set_target_properties(${CLIENT_TARGET}
 		PROPERTIES CXX_STANDARD 17
